requirements.txt
playwright
requests
beautifulsoup4
rapidfuzz
pdfminer.six
